+++ packages/article-list/src/article-list-item-separator.js
@@ -1,0 +1,9 @@
+import React from "react";
+import { View } from "react-native";
+import styles from "./styles";
+
+const ArticleListItemSeparator = () => (
+  <View style={styles.listItemSeparator} />
+);
+
+export default ArticleListItemSeparator;