+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.10"></a>
+## [0.3.10](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.3.9...@times-components/video-label@0.3.10) (2018-05-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
 <a name="0.3.9"></a>
 ## [0.3.9](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.3.8...@times-components/video-label@0.3.9) (2018-05-08)